Module arc_runtime::prelude

Re-exports

pub use crate::context::Context;

pub use crate::control::Control;

pub use crate::control::Control::Continue;

pub use crate::control::Control::Finished;

pub use crate::data::channels;

pub use crate::data::channels::Channel;

pub use crate::data::garbage::Alloc;

pub use crate::data::garbage::Gc;

pub use crate::data::primitives::bool;

pub use crate::data::primitives::char;

pub use crate::data::primitives::f32;

pub use crate::data::primitives::f64;

pub use crate::data::primitives::i128;

pub use crate::data::primitives::i16;

pub use crate::data::primitives::i32;

pub use crate::data::primitives::i64;

pub use crate::data::primitives::i8;

pub use crate::data::primitives::u128;

pub use crate::data::primitives::u16;

pub use crate::data::primitives::u32;

pub use crate::data::primitives::u64;

pub use crate::data::primitives::u8;

pub use crate::data::primitives::unit;

pub use crate::data::primitives::unit;

pub use crate::data::primitives::unit;

pub use crate::data::primitives::unit;

pub use crate::data::primitives::Unit;

pub use crate::data::primitives::Unit;

pub use crate::data::primitives::Unit;

pub use crate::data::primitives::Unit;

pub use crate::data::Abstract;

pub use crate::data::Concrete;

pub use crate::data::DynSendable;

pub use crate::data::DynSharable;

pub use crate::data::Sendable;

pub use crate::data::Sharable;

pub use crate::runtime::Runtime;

pub use crate::task::message::TaskMessage;

pub use crate::data::primitives::assert;

pub use crate::data::primitives::panic;

pub use crate::data::primitives::print;

pub use crate::data::strings::String;

pub use crate::data::strings::String_clear;

pub use crate::data::strings::String_from_str;

pub use crate::data::strings::String_insert_char;

pub use crate::data::strings::String_is_empty;

pub use crate::data::strings::String_len;

pub use crate::data::strings::String_new;

pub use crate::data::strings::String_push_char;

pub use crate::data::strings::String_push_str;

pub use crate::data::strings::String_remove;

pub use crate::data::strings::String_split_off;

pub use crate::data::strings::String_with_capacity;

pub use crate::data::strings::String_from_i32;

pub use crate::data::strings::String_eq;

pub use crate::data::strings::String_concat;

pub use crate::data::vectors::Vec;

pub use crate::data::vectors::Vec_capacity;

pub use crate::data::vectors::Vec_clear;

pub use crate::data::vectors::Vec_dedup;

pub use crate::data::vectors::Vec_get;

pub use crate::data::vectors::Vec_insert;

pub use crate::data::vectors::Vec_is_empty;

pub use crate::data::vectors::Vec_len;

pub use crate::data::vectors::Vec_new;

pub use crate::data::vectors::Vec_pop;

pub use crate::data::vectors::Vec_push;

pub use crate::data::vectors::Vec_remove;

pub use crate::data::vectors::Vec_with_capacity;

pub use crate::data::cells::Cell;

pub use crate::data::cells::Cell_get;

pub use crate::data::cells::Cell_new;

pub use crate::data::cells::Cell_set;

pub use std::task::Poll::Pending;

pub use std::task::Poll::Ready;

Macros

Debug

Derive macro generating an impl of the trait Debug.

Hash

Derive macro generating an impl of the trait Hash.

_vector

access

Access a struct’s field.

call

call_async

call_indirect

convert_reflexive

date

declare

declare_functions

enwrap

Enwraps a value into an enum-variant.

erase

function

hexf32

Expands to a f32 value with given hexadecimal representation.

hexf64

Expands to a f64 value with given hexadecimal representation.

info

Log info level record

is

Returns true if enum is a certain variant, else false.

letroot

new

Constructs a struct.

pull

pull_transition

push

push_transition

terminate

time

transition

unerase

unwrap

Unwraps a value out of an enum-variant.

val

Get the value of a variable.

vector

wait

warn

Log warning level record

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

Component

A concrete component instance

ComponentContext

The contextual object for a Kompact component

ComponentDefinitionAccess

Gives access to the component definition within a future/async function

DateTime

Combined date and time.

DeadletterBox

The default deadletter box

Duration

A span of time with nanosecond precision.

ExecuteResult

Statistics about the last invocation of execute.

Immix

Immix GC implementation. Read top level module documentation for more information

ImmixOptions

KompactConfig

A configuration builder for Kompact systems

KompactSystem

A Kompact system is a collection of components and services

MutatorRef

Reference counted Mutator instance. When there is no references to this mutator mutator is detached from GC heap.

NetMessage

An incoming message from the networking subsystem

NetworkConfig

Configuration builder for the network dispatcher

Pin

A pinned pointer.

PollContext

The context of an asynchronous task.

Rooted

Rooted value on stack. This is non-copyable type that is used to hold GC thing on stack.

ShadowStack

Shadow stack implementation. Internally this is singly-linked list of on stack rooted values.

ShadowStackInternal

Almost the same as raw entry of shadow stack except this one gives access to value. This type is not exposed in public API and used only internally.

TypeId

A TypeId represents a globally unique identifier for a type.

UnsafeCell

The core primitive for interior mutability in Rust.

Enums

AllocationSpace

Handled

State transition indication at the end of a message or event handler

MsgEnvelope

A message that is accepted by an actor’s mailbox

Poll

Indicates whether a value is available or if the current task has been scheduled to receive a wakeup instead.

Traits

Actor

A slightly higher level Actor API that handles both local and networked messages

ActorRaw

The base trait for all actors

Any

A trait to emulate dynamic typing.

Collectable

Indicates type that can be allocated on garbage collector heap.

ComponentDefinition

The core trait every component must implement

ComponentLifecycle

A trait to customise handling of lifecycle events

ComponentLogging

An abstraction over providers of Kompact loggers

Debug

? formatting.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

Deserializer

A data format that can deserialize any data structure supported by Serde.

DynamicPortAccess

A mechanism for dynamically getting references to provided/required ports from a component.

Finalize

Indicates a type that can be finalized after marking phase.

Future

A future represents an asynchronous computation obtained by use of async.

FutureExt

An extension trait for Futures that provides a variety of convenient adapters.

Hash

A hashable type.

Rootable

Trait that should be implemented for all types that could be rooted. In simple cases impl<T: Traceable> Rootable for T {} is enough.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Serializer

A data format that can serialize any data structure supported by Serde.

SystemHandle

A limited version of a KompactSystem

Trace

Indicates that a type can be traced by a garbage collector.

Visitor

Functions

block_on

Run a future to completion on the current thread.

instantiate_immix

replace_with_or_abort_and_return

Temporarily takes ownership of a value at a mutable location, and replace it with a new value based on the old one. Aborts on panic. Lets the closure return a custom value as well.

Type Definitions

BoxFuture

An owned dynamically typed Future for use in cases where you can’t statically type your result or need to add some indirection.

Never

A more readable placeholder for a stable Never (!) type.

Attribute Macros

rewrite

Derive Macros

Abstract

Actor

A macro to derive empty actor implementations

Alloc

Collectable

ComponentDefinition

A macro to derive fair ComponentDefinition implementations

Deserialize

Finalize

NoDebug

NoSerde

NoTrace

Send

Serialize

Sync

Trace

Unpin