Struct arc_runtime::prelude::Component

pub struct Component<CD>where
    CD: ComponentTraits,{
    core: ComponentCore,
    custom_scheduler: Option<DedicatedThreadScheduler>,
    pub(crate) mutable_core: Mutex<ComponentMutableCore<CD>>,
    ctrl_queue: SegQueue<ControlEvent>,
    msg_queue: TypedMsgQueue<<CD as ActorRaw>::Message>,
    supervisor: Option<ProvidedRef<SupervisionPort>>,
    logger: Logger<Arc<Fuse<Async>>>,
    recovery_function: Mutex<Box<dyn FnOnce(FaultContext) -> RecoveryHandler + Send + 'static, Global>>,
}

A concrete component instance

The component class itself is application agnostic, but it contains the application specific ComponentDefinition.

Fields
core: ComponentCorecustom_scheduler: Option<DedicatedThreadScheduler>mutable_core: Mutex<ComponentMutableCore<CD>>ctrl_queue: SegQueue<ControlEvent>msg_queue: TypedMsgQueue<<CD as ActorRaw>::Message>supervisor: Option<ProvidedRef<SupervisionPort>>logger: Logger<Arc<Fuse<Async>>>recovery_function: Mutex<Box<dyn FnOnce(FaultContext) -> RecoveryHandler + Send + 'static, Global>>
Implementations

impl<CD> Component<CD>where CD: ComponentTraits,

pub fn definition_mut(&mut self) -> &mut CD

Returns a mutable reference to the underlying component definition.

This can only be done if you have a reference to the component instance that isn’t hidden behind an Arc. For example, after the system shuts down and your code holds on to the last reference to a component you can use get_mut or try_unwrap.

pub fn on_definition<T, F>(&self, f: F) -> Twhere F: FnOnce(&mut CD) -> T,

Execute a function on the underlying ComponentDefinition and return the result

This method will attempt to lock the mutex, and then apply f to the component definition inside the guard.

pub fn logger(&self) -> &Logger<Arc<Fuse<Async>>>

Returns a reference to this component’s logger

pub fn is_faulty(&self) -> bool

Returns true if the component is marked as faulty.

pub fn is_active(&self) -> bool

Returns true if the component is marked as active.

pub fn is_destroyed(&self) -> bool

Returns true if the component is marked as destroyed.

pub fn wait_ended(&self)

Wait synchronously for this component be either destroyed or faulty

This component blocks the current thread and hot-waits for the component to become either faulty or destroyed. It is meant mostly for testing and not recommended in production.

pub fn set_recovery_function<F>(&self, f: F)where F: FnOnce(FaultContext) -> RecoveryHandler + Send + 'static,

Set the recovery function for this component

See RecoveryHandler for more information.