Struct arc_runtime::prelude::DeadletterBox

pub struct DeadletterBox {
    ctx: ComponentContext<DeadletterBox>,
    notify_ready: Option<KPromise<()>>,
}

The default deadletter box

Simply logs every received message at the info level and then discards it.

Fields
ctx: ComponentContext<DeadletterBox>notify_ready: Option<KPromise<()>>
Implementations

impl DeadletterBox

pub fn new(notify_ready: KPromise<()>) -> DeadletterBox

Creates a new deadletter box

The notify_ready promise will be fulfilled, when the component received a Start event.