Struct arc_runtime::prelude::KompactConfig

pub struct KompactConfig {
    pub(crate) label: String,
    pub(crate) throughput: usize,
    pub(crate) msg_priority: f32,
    pub(crate) threads: usize,
    pub(crate) timer_builder: Rc<dyn Fn() -> Box<dyn TimerComponent + 'static, Global> + 'static>,
    pub(crate) scheduler_builder: Rc<dyn Fn(usize) -> Box<dyn Scheduler + 'static, Global> + 'static>,
    pub(crate) sc_builder: Rc<dyn Fn(&KompactSystem, KPromise<()>, KPromise<()>) -> Box<dyn SystemComponents + 'static, Global> + 'static>,
    pub(crate) root_logger: Option<Logger<Arc<Fuse<Async>>>>,
    pub(crate) config_sources: Vec<ConfigSource, Global>,
}

A configuration builder for Kompact systems

Example

Set a custom label, and run up to 50 events or messages per scheduling of a component on a threadpool with 2 threads.

use kompact::prelude::*;

let mut conf = KompactConfig::default();
conf.label("My special system")
    .throughput(50)
    .threads(2);
let system = conf.build().expect("system");

Fields
label: Stringthroughput: usizemsg_priority: f32threads: usizetimer_builder: Rc<dyn Fn() -> Box<dyn TimerComponent + 'static, Global> + 'static>scheduler_builder: Rc<dyn Fn(usize) -> Box<dyn Scheduler + 'static, Global> + 'static>sc_builder: Rc<dyn Fn(&KompactSystem, KPromise<()>, KPromise<()>) -> Box<dyn SystemComponents + 'static, Global> + 'static>root_logger: Option<Logger<Arc<Fuse<Async>>>>config_sources: Vec<ConfigSource, Global>
Implementations

impl KompactConfig

pub fn new() -> KompactConfig

👎Deprecated since 0.11.0: If you really want these exact parameters, use KompactConfig::load_config_str(kompact::runtime::MINIMAL_CONFIG) instead. Otherwise prefer KompactConfig::default().

Create a minimal Kompact config

The minimal config uses the default label, i.e. "kompact-runtime-{}" for some sequence number.

It sets the event/message throughput to 2, split evenly between events and messages.

It runs with a single thread on a small pool.

It uses all default components, without networking, with the default timer and default logger.

pub fn label<I>(&mut self, s: I) -> &mut KompactConfigwhere I: Into<String>,

👎Deprecated since 0.11.0: Use KompactConfig::set_config_value(&kompact::config_keys::system::LABEL, ...) instead.

Set the name of the system

The label is used as metadata for logging output.

pub fn throughput(&mut self, n: usize) -> &mut KompactConfig

👎Deprecated since 0.11.0: Use KompactConfig::set_config_value(&kompact::config_keys::system::THROUGHPUT, ...) instead.

Set the maximum number of events/messages to handle before rescheduling a component

Larger values can increase throughput on highly loaded components, but at the cost of fairness between components.

pub fn msg_priority(&mut self, r: f32) -> &mut KompactConfig

👎Deprecated since 0.11.0: Use KompactConfig::set_config_value(&kompact::config_keys::system::MESSAGE_PRIORITY, ...) instead.

Set the ratio between handling messages and events.

A component will handle up to throughput * r messages and throughput * (1-r) events before rescheduling.

If there are less events or messages than alloted queued up the remaining allotment will be redistributed to the other type until all throughput is used up or no messages or events remain.

pub fn threads(&mut self, n: usize) -> &mut KompactConfig

👎Deprecated since 0.11.0: Use KompactConfig::set_config_value(&kompact::config_keys::system::THREADS, ...) instead.

The number of threads in the Kompact thread pool

Note

You must ensure that the selected scheduler implementation can manage the given number of threads, if you customise this value!

pub fn scheduler<F>(&mut self, f: F) -> &mut KompactConfigwhere F: Fn(usize) -> Box<dyn Scheduler + 'static, Global> + 'static,

Set a particular scheduler implementation

Takes a function f from the number of threads to a concrete scheduler implementation as argument.

pub fn executor<E, F>(&mut self, f: F) -> &mut KompactConfigwhere E: FuturesExecutor + Sync + 'static, F: Fn(usize) -> E + 'static,

Set a particular scheduler implementation based on a FuturesExecutor

Takes a function f from the number of threads to a concrete executor as argument.

pub fn timer<T, F>(&mut self, f: F) -> &mut KompactConfigwhere T: TimerComponent + 'static, F: Fn() -> Box<dyn TimerComponent + 'static, Global> + 'static,

Set a particular timer implementation

pub fn system_components<B, C, FB, FC>( &mut self, deadletter_fn: FB, dispatcher_fn: FC ) -> &mut KompactConfigwhere B: ComponentDefinition<Message = !> + ActorRaw + 'static, C: ComponentDefinition<Message = DispatchEnvelope> + ActorRaw + 'static + Dispatcher, FB: Fn(KPromise<()>) -> B + 'static, FC: Fn(KPromise<()>) -> C + 'static,

Set a particular set of system components

In particular, this allows exchanging the default dispatcher for the NetworkDispatcher, which enables the created Kompact system to perform network communication.

Example

For using the network dispatcher, with the default deadletter box:

use kompact::prelude::*;

let mut cfg = KompactConfig::new();
cfg.system_components(DeadletterBox::new, {
    let net_config = NetworkConfig::new("127.0.0.1:0".parse().expect("Address should work"));
    net_config.build()
});
let system = cfg.build().expect("KompactSystem");

pub fn system_components_with_dedicated_dispatcher<B, C, FB, FC>( &mut self, deadletter_fn: FB, dispatcher_fn: FC ) -> &mut KompactConfigwhere B: ComponentDefinition<Message = !> + ActorRaw + 'static, C: ComponentDefinition<Message = DispatchEnvelope> + ActorRaw + 'static + Dispatcher, FB: Fn(KPromise<()>) -> B + 'static, FC: Fn(KPromise<()>) -> C + 'static,

Set a particular set of system components

This function works just like system_components, except that it assigns the dispatcher to its own thread using create_dedicated_unsupervised.

pub fn logger(&mut self, logger: Logger<Arc<Fuse<Async>>>) -> &mut KompactConfig

Set the logger implementation to use

pub fn load_config_file<P>(&mut self, path: P) -> &mut KompactConfigwhere P: Into<PathBuf>,

Load a HOCON config from a file at path

This method can be called multiple times, and the resulting configurations will be merged.

It matters in which order configs are loaded and values are set. See HoconLoader for more information.

The loaded config can be accessed via [system.config()](KompactSystem::config or from within a component via [self.ctx().config()](ComponentContext::config.

pub fn load_config_str<S>(&mut self, config: S) -> &mut KompactConfigwhere S: Into<String>,

Load a HOCON config from a string

This method can be called multiple times, and the resulting configurations will be merged.

It matters in which order configs are loaded and values are set. See HoconLoader for more information.

The loaded config can be accessed via [system.config()](KompactSystem::config or from within a component via [self.ctx().config()](ComponentContext::config.

pub fn set_config_value<T>( &mut self, config: &ConfigEntry<T>, value: <T as ConfigValueType>::Value ) -> &mut KompactConfigwhere T: ConfigValueType,

Override a single value in the HOCON config

This method can be called multiple times and the resulting configurations will be merged.

It matters in which order configs are loaded and values are set. See HoconLoader for more information.

pub fn build(self) -> Result<KompactSystem, KompactError>

Finalise the config and use it create a KompactSystem

This function can fail, if the configuration sets up invalid schedulers or dispatchers, for example.

Example

Build a system with default settings with:

use kompact::prelude::*;

let system = KompactConfig::default().build().expect("system");