Trait arc_runtime::prelude::Collectable

pub trait Collectable: Trace + Finalize + Any {
    // Provided method
    fn allocation_size(&self) -> usize { ... }
}

Indicates type that can be allocated on garbage collector heap.

Provided Methods

fn allocation_size(&self) -> usize

Function to compute value size on allocation. If type is dyn sized (i.e array or string) this function must be overloaded to calculate allocation size properly. It is invoked exactly once at allocation time.

Implementations

impl dyn Collectable + 'static

pub fn is<T>(&self) -> boolwhere T: Collectable,

Returns true if the boxed type is the same as T

pub fn downcast_ref<T>(&self) -> Option<&T>where T: Collectable,

Returns some reference to the boxed value if it is of type T, or None if it isn’t.

pub unsafe fn downcast_ref_unchecked<T>(&self) -> &Twhere T: Collectable,

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T>where T: Collectable,

Returns some mutable reference to the boxed value if it is of type T, or None if it isn’t.

pub unsafe fn downcast_mut_unchecked<T>(&mut self) -> &mut Twhere T: Collectable,

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl dyn Collectable + 'static

pub fn downcast<T>( self: Box<dyn Collectable + 'static, Global> ) -> Result<Box<T, Global>, Box<dyn Collectable + 'static, Global>>where T: Collectable,

Returns the boxed value if it is of type T, or Err(Self) if it isn’t.

pub unsafe fn downcast_unchecked<T>( self: Box<dyn Collectable + 'static, Global> ) -> Box<T, Global>where T: Collectable,

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.