Trait arc_runtime::prelude::ComponentLifecycle

pub trait ComponentLifecycle: ComponentLogging {
    // Provided methods
    fn on_start(&mut self) -> Handled
       where Self: 'static { ... }
    fn on_stop(&mut self) -> Handled
       where Self: 'static { ... }
    fn on_kill(&mut self) -> Handled
       where Self: 'static { ... }
}

A trait to customise handling of lifecycle events

This trait replaces the pre-v0.9 Provide<ControlPort> requirement.

Provided Methods

fn on_start(&mut self) -> Handledwhere Self: 'static,

Gets invoked every time a component receives a Start event

The default implementation simply logs something at debug level.

fn on_stop(&mut self) -> Handledwhere Self: 'static,

Gets invoked every time a component receives a Stop event

The default implementation simply logs something at debug level.

fn on_kill(&mut self) -> Handledwhere Self: 'static,

Gets invoked every time a component receives a Kill event

The default implementation simply logs something at debug level.