Trait arc_runtime::prelude::Trace

pub unsafe trait Trace {
    // Provided method
    fn trace(&mut self, _vis: &mut dyn Visitor) { ... }
}

Indicates that a type can be traced by a garbage collector.

This doesn’t necessarily mean that the type is safe to allocate in a garbage collector (Collectable).

Safety

See the documentation of the trace method for more info. Essentially, this object must faithfully trace anything that could contain garbage collected pointers or other Trace items.

Provided Methods

fn trace(&mut self, _vis: &mut dyn Visitor)

Trace each field in this type.

Structures should trace each of their fields, and collections should trace each of their elements.

Safety

Some types (like Gc) need special actions taken when they’re traced, but those are somewhat rare and are usually already provided by the garbage collector.

Behavior is restricted during tracing:

Permitted Behavior

Reading your own memory (includes iteration)

Interior mutation is undefined behavior, even if you use RefCell

Calling Visitor::mark_object

Panicking on unrecoverable errors

This should be reserved for cases where you are seriously screwed up, and can’t fulfill your contract to trace your interior properly.

One example is Gc<T> which panics if the garbage collectors are mismatched

Garbage collectors may chose to abort if they encounter a panic, so you should avoid doing it if possible.

Never Permitted Behavior

Forgetting a element of a collection, or field of a structure

If you forget an element undefined behavior will result

This is why you should always prefer automatically derived implementations where possible.

With an automatically derived implementation you will never miss a field

It is undefined behavior to mutate any of your own data.

The mutable &mut self is just so copying collectors can relocate GC pointers

Calling other operations on the garbage collector (including allocations)